pyproject.toml
[project]
name = "recsys"
version = "0.1.0"
description = "E-commerce Recommendation System for Master's Thesis"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Author" }]

dependencies = [
    # Data processing
    "polars>=1.0.0",
    "duckdb>=1.0.0",
    "pyarrow>=17.0.0",

    # Machine Learning
    "implicit>=0.7.0",
    # "lightfm>=1.17",  # requires C compiler, install separately
    "mlxtend>=0.23.0",
    "scikit-learn>=1.5.0",
    "catboost>=1.2.0",

    # API & UI
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "streamlit>=1.38.0",

    # MLOps & Visualization
    "mlflow>=2.16.0",
    "plotly>=5.24.0",

    # Utils
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
